Move Track Requests state into separate Vuex module
as quick example of modularity in action here.  (This turned out to be significantly easier than moving the state for active study requests, especially now that we have comments and dependent fetches for users and location.)
